backend-python/ai_server/requirements.txt
# Core dependencies
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# AI and LLM
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0

# Database
supabase==2.3.0

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Redis for caching and rate limiting
redis==5.0.1
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# HTTP client
httpx>=0.24.0,<0.25.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
isort==5.13.2
flake8==7.0.0